Task methods
Added the set_completed methods (amongst others)
Added some smart checks that will check to see if any two of duration,
completed, or begin times are set when one of the three are updated to
fill in the other one accordingly. This one of them could be gotten rid
of and then turn this into a calculation but then it would make it tough
to save any combination that the user prefers.