Not trying angular again, uploading to heroku
Below the line is what we were working with in angular. we decided to
not go this route because the documentation doesn't seem that great and
there were a lot of complaints about it. We could be completely wrong
but backbone is more familiar.

****************************************************************************

the guide that I as using just seems to not work so I'm looking up
another one.

http://mightygio.com/2013/03/integrating-rails-and-angularjs-part-1-structure/

--Added the javascripts\config folder as reccomended
--Added the constants / debug files
--Added csrf.js.coffee which