wFinally herokuw swhould workwwwwwwwwwwwwwwwwwww
After searching around for 45minutes I couldn't find a way to get
subtrees to work in windows so I'm moving everything out of the api
folder because I just want this to work and im tired of messing with it
for something so small as a subfolder. and my w key slowly dwwying hich
is aggwwrawvatwing me. lets hope this worksw